+++ src/pages/ReportPage.tsx
@@ -1,6 +1,6 @@
-import CalendarHeader from '@ui/components/layout/CalendarHeader';
+import CalendarHeader from '@ui/components/layout/calendar/CalendarHeader';
 import Body from '@ui/components/layout/Body';
-import Menu from '@ui/components/layout/Menu';
+import Menu from '@ui/components/layout/MenuBar/Menu';
 
 export default function ReportPage() {
   return (
