+++ src/pages/SettingPage.tsx
@@ -1,5 +1,5 @@
 import Body from '@ui/components/layout/Body';
-import Menu from '@ui/components/layout/Menu';
+import Menu from '@ui/components/layout/MenuBar/Menu';
 import SignButton from '@ui/components/layout/common/SignButton';
 import styleTokenCss from '@ui/styles/styleToken.css';
 import styled from '@emotion/styled';
